requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pillow==10.1.0